Added 2 pieces of functionality, a Team Model line and the ability to print using a secondary function column.
The Team Model is used by defining TEAM_MODEL in the appropriate class in orgchart_parser, lining up the key in the map with the name of the product team.

The secondary function column is used by passing in a -a flag to ppt_draw, and adding an __init__ to the class in orgchart_parser that will select the appropriate secondary column to display.  This is used so that you can have two views of the org, the actual current view, and the model (i.e., we have temporarily overloaded someone into Automation even though in the model that head count would go to QA).